SwiftDriver: initial work to properly handle android cross-compilation
The intent here is to permit the Windows/macOS style cross-compilation
for Android. This involves passing `-sdk` with the path to the "Swift
SDK" which overlays the system's native SDK (NDK). The
`ANDROID_NDK_ROOT` is a well-defined environment variable (setup by the
SDK installer as well as a general expectation for Android development)
that identifies the root of the installation of the NDK. This allows us
to locate the native SDK root (`--sysroot`) for driving the linker
driver amongst other paths.