[Linux] Add build-id support.
Add a `-build-id` option to the driver to let users control the build
ID setting used by the linker.  Pass `--build-id` to the linker by default,
so that build IDs are enabled by default but using the linker's default
type; if someone wishes to specify a particular option, they can do so
using the new driver option.

Also turn on build IDs when *building* the driver.

rdar://116798309